Fix bugs exposed by tests.
The big bug is that I was misusing pointers in btrfs.Pull, which lead to
the same commit getting sent again and again. Bugs like this make me
wish I actually knew how to program.

The smaller bug was that I was that I wasn't making replica repos for
the replica shards in tests.