Update makefile to include command to easily test all local unit tests. I've used the 'short' flag to delineate between integration tests and tests that can run locally. There are a few more that I need to flag this way / or have them check for the existence of and environment variable to decide if they can run.
Also - I had to update a bunch of calls to commit.Id to commit.ID for linting ... but this is silly
because its very inconsistent since the pb go objects are created w fields that are listed as 'Id'
(and these are autogenerated). So .... it'll be super inconsistent if we don't just use 'Id'
everywhere. I guess I've re-derived why they were named like this.

Instead, if we're going to use the linter, maybe we can omit errors of that type somehow.