Relocates TestHandleRace test.
Unfortunately this loses us a bit of test coverage, because we're
testing handles through the go api rather than the fuse layer. Testing
this through the fuse layer would have been very difficult though
because it would require making the test helper mount the repo twice. It
didn't seem worth it just for this one test although we should do it in
the future.