Fix edge case w sampling probablities
Numpy was complaining that the probability distribution vector didn't
sum to 1. In fact it was very close. This seems to be a bug w the numpy
sampling function:

https://github.com/sherjilozair/char-rnn-tensorflow/issues/1

So I re-normalized the vector just in case, did the random sampling
myself, and in the case the sampling hits the 'hole' (the low
        probability because its not close to 1), it selects the item w
the highest probability.