Start converting all the key helper calls over to the new method
Ran into an interesting case in getDiffsInCommitRange. This takes its own key generator
helper function because its called with different indices. I'll need to update that caller
to be able to handle the helper returning an error. It may also make sense to flatten this
out a bit and make a helper per index once again (just using this new method).