Fix a bug where output commits are not parented correctly.  That was because when a job finishes, it tries to use the output commit of its parent job as the parent commit of its own output commit.  However, the output commit of the parent job may not have been started yet, because the parent job may not have finished.
So the solution is to change how Merge works.  Basically Merge now
takes an open commit as an input.  So when a job starts, we start
a commit and use it as the output commit.  Then when the job finishes
we use the output commit for Merge.  That way, even if a job's parent
has not finished, it can still find the parent's output commit.