Lazily dial the block server within driver
This fixes an issue in the /src/server/pfs/server test suite. That test harness
initializes the driver (which was immediately dialing the block server), THEN
creating/registering the block server. This change seems to fix the test harness.

HOWEVER - whats odd is that pachd's init function does almost the exact same
thing (initializing the driver before registering the block server). So, our
best guess is that GRPC changed how it handles reconnecting ... and backs off
more aggressively. And in the case of the test suite it means that the conn
isn't avail in time, but in the case of pachd it works OK. Very odd