size can be 0 for Reader() by default, handle it.
In minioClient pfs driver implementation Reader()
did not handle the case of reading the entire file
i.e when size was provided as '0'.

This is the default behavior for someone reading a file

```
pachctl get-file testminio master /blah
```

Re-factor the code a bit to address the cases of offset and
size independently without using the SectionReader() since
it is not guaranteed that size is always > 0.

Fixes #1385