server/pkg/worker: update datum hashing algorithm
Currently, we hash a datum using files + transform.  The consequence is
that if a pipeline has been updated without updating its transform, data
is not re-processed, which is not-ideal.

The new hashing algorithm takes pipeline name, ID, and version into
account, to ensure that if a pipeline is updated/re-created, we
re-process the input data.