server/pkg/sync: fix deadlocks within lazy pulling
This patch fixes a number of issues with lazy pulling that could result
in deadlocks.  Below is the list of issues.  We assume that the
concurency limit is N and the number of files is M, where M > N.

1. If a process opens one of the `M-N` files, the process blocks.
2. After the CleanUp goroutine runs, up to `M-N` goros might end up
blocking on a syscall, thus resulting in `M-N` deadlocked OS threads.
3. After CleanUp runs, up to N goros might proceed to download files
anyways, even if no one will be using the files anymore.