Update documentation for Google Cloud deployments
When I first tried to deploy Pachyderm on Google Cloud, I ran into two problems:
 * Insufficient Permission errors when adding data to a repository, and
 * No dashboard at localhost:... when I ran pachctl port-forward.

The first problem was caused by creating the Kubernetes cluster via the Cloud console, which I thought would be easier than using the command-line tool. It's important to use the gcloud tool because scopes can't be specified in the console.

The second problem was caused by not including the --dashboard flag when installing Pachyderm on the cluster.