Fix bug & add a corner-case test for DeleteCommit
This adds a test for the case where several consecutive commits all
share a single upstream commit in their provenance (e.g. a pipeline with
two inputs, one with many commits, and another with one or two) and then
a high out-degree commit is deleted. This revealed a bug in the
branch-update algorithm in DeleteCommit, which this also fixes. The bug
is described in the comments of pachyderm/pachyderm/pull/2618