Remove GitHubUsername from Authenticate & Activate
Instead do one of the following two things:

1) In dev mode (pachctl deploy local):
Try to detect whether the token that the caller has passed is a GitHub
Access Code, and if not, automatically authenticate them as the GitHub
user whose name is equal to the access code that they provided (making
it easy to authenticate as any GitHub user for testing, and avoiding any
special-case code in the Pachyderm Dashboard).

2) In non-dev mode:
Always infer the caller's username from the GitHub token that they
provided. As Gabriel pointed out, any user who has one of these codes
can convert it to a Username by sending it to GitHub, so there's no
reason to try and "verify" that a given GitHub user has the right code
for themselves.