pachctl list-job filters, but list-job -p auths
Previously, list-job would check if the caller had access to the output
repo of every job that it returns, and if the auth check failed on any
repo, list-job would return an error. This meant that if a user didn't
have access to any output repo in a cluster, `list-job` would fail.

Now, if a user doesn't have access to an output repo in a cluster,
list-job simply doesn't return jobs that output to that repo. If a user
explicitly specifies a pipeline such that the user doesn't have access
to the pipeline's output repo, then list-job returns an error.