Update Makefile with correct pachctl path
Fixing an issue that calls the correct pachctl to get client version. The change here restores the behavior of Makefile when VERSION file was used to set the VERSION env variable passed to various build scripts.

By explicitly refering to $(GOPATH)/bin the version become independent of developers PATH varaible.