Create pipeline info transactionally (#5431)
The main obstacles in transactionifying pipeline creation are finishing open output commits and creating the pipeline info spec commit. Other steps (branch and non-spec commit creation) can be safely handled by existing STM logic.

Now that the output branch always has provenance, finishing the output commits is no longer a correctness issue. To reduce the possibility of missing output commits created during the transaction, this step is deferred until right before commit provenance propagation. 

To hold the info file, we need to create a commit outside of the transaction. The tentative commit we create is kept around as long as possible, only deleted if the transaction is or if there's an error. Pipeline creation includes some additional validation checks to ensure no updates have occurred in the mean time. For now, though, the transaction just gives up if a change is encountered rather than trying to recreate, due to some challenges around modifying a saved transaction.