Widen retry logic in PPS master.
In some cases, a "step" operation failing to complete can indicate that
the underlying pipeline state has changed, making the step obsolete.
Instead of myopically retrying the same operation and eventually
failing, retrying the step from the beginning provides more confidence
that PPS master is acting based on up-to-date information.