Attempt to parse combined stats commits via BlockRefs.
When a datum fails, jobs still produce output to be included in the
stats commit. This violates the assumptions made while processing
additive jobs, leading to new stats info being combined with any files
already in the parent commit. Hashtrees are merged in a predictable way,
so we can attempt to perform a diff against the parent looking only at
the list of block references, and recover the proper file bytes.

Also stop returning datum info for skipped datums, which rely on out of
date datum indices and are therefore unreliable.