Fix beginner tutorial warning (#5705)
* Prevent warning from pachctl when creating montage pipeline

The existing `montage.json` uses the image `v4tech/imagemagick`, with
no tag indicated.  This results in a warning:

```
WARNING: please specify a tag for the docker image in your transform.image spec.
For example, change 'python' to 'python:3' or 'bash' to 'bash:5'. This improves
reproducibility of your pipelines.
```

Using the `latest` tag results in a different warning:

```
WARNING: please do not specify the ':latest' tag for the docker image in your
transform.image spec. For example, change 'python:latest' to 'python:3' or
'bash:latest' to 'bash:5'. This improves reproducibility of your
pipelines.
```

The `v4tech/imagemagick` has no other tags; also, it has not been
updated in four years.

There is an alternate repo `dpokidov/imagemagick`; I tested
`dpokidov/imagemagick:7.0.10-58` and it generates the same montage
image with the example data.

* Fix typo

Missing ‘/.’