Keeps Object Storage and Postgres in Sync (#5818)
These changes let us create and delete chunks without ever ending up with a chunk without a corresponding object in object storage. It is possible to wind up with multiple objects per chunk, and objects without a Postgres entry, but the existence of these anomalies will not impact correctness, and there are background services to eliminate them eventually.