Adds a table for secrets in Postgres. (#5893)
We currently use convergent encryption to encrypt the chunks. This allows us to de-duplicate same as without encryption using content-defined chunking, and content-addressed storage. A given plaintext will always produce the same ciphertext. One downside to this, is that someone with the same data as us can check if we have stored the data. These changes add a secret for the cluster which is used to derive keys for convergent encryption. The effect is a single cluster produces the same ciphertexts for a given plaintext, but different clusters will not produce the same ciphertexts.
- There is no new cryptography in this P.R. It simply wires up a field that was already there.
- Adds a table for storing secret keys
- During the PFS driver startup: we create a default key if it does not exist
- Also adds an encryption_algo field to the chunk refs.