De-flake TestCancelManyJobs
This test creates a pipeline and 10 input commits, spawning 10 jobs, and
then attempts to cancel each job, one-by-one, asserting that
cancellation is always successful (the job comes up and the StopJob()
RPC succeeds)

The issue here is that the datum the pipeline uses (/file) is not
modified in any of the input commits that this test creates. Jobs,
therefore, skip this datum and finish before the test can cancel them,
failing the test's assertion that the job comes up, and also causing the
StopJob() RPC to fail (as the job is already finished).

I'm not sure why this wasn't an issue in 1.x, but maybe datum skipping
is faster in 2.0, such that previously the test could assert that each
job is starting and cancel it before the job itself is able to skip its
input datum, but now skipping happens faster than the test is able to
cancel the job?