Ensure /tmp, /pach-cache exist in the pachd image. (#6065)
We attempt to write to /tmp in multiple places, which is fine if we're
running on a normal image with a writable /tmp, or are running as root.
The pachd image is built 'FROM scratch', so to avoid having to be root,
we just make sure /tmp exists. /pach-cache is the default value of
PACH_CACHE_ROOT, which should probably work by default if not root.

There's still a potential issue with non-root for local deployments: we
try to use host node storage for the object store and etcd, which won't
be writable by default. We regard local deployments as toy/experimental
setups, though, so this was deemed not a problem.

Also move the other cache under CACHE_ROOT, though this is no longer
necessary.