Eliminate most uses of Repo.Name (#6164)
System repos break the assumption that a repo can be fully specified by its name, which is used quite often in our code base. Some of this is caused by function signatures requiring PFS entities to be specified with strings, leading to fields being taken out of a proto to call the function. To avoid unforeseen issues cropping up as system repos are implemented, I've removed the vast majority of direct uses of the name field of a repo. The main exceptions are related to CLI output.

This PR does not touch the fuse or s3 code, which use repo names to specify other objects (e.g. buckets), though these will need to be addressed as well.