Allow modifications to transactions while finishing.
Creating pipelines in a transaction has the dubious distinction that we
expect it to fail and update its saved response when it encounters
modifications outside of the transaction. To accommodate this logic,
appending a CreatePipelineRequest to a transaction involves updating the
stored transaction info to reflect the most recent state of the
pipeline's spec repo.

However, another source of inside/outside transaction conflicts is
during FinishTransaction, after the append logic has finished.
Previously, the finishing code was not able to modify the stored
transaction, ended up looping in a futile attempt to resolve the
conflict.

The new logic may be overengineered, but provides a straightforward way
to modify in a transaction in a way that should be persisted in some
cases.

Also, modified the error for a transaction being modified to have a name
better reflecting that scenario, as opposed to one almost identical to
an unrelated SQL-level error.