Allow moving a branch to an ancestor of its head. (#6465)
CreateBranch, when called as a standalone command and used to move a branch's head, attempts to use the target commit's commit set. This may not be possible if there are downstream commits in that commit set already, and could lead to incorrect results if the provenance graph has changed since that commit set. To avoid part of this problem, we are more strict about when an old alias can be reused. In that error case, or others that point to trying to include multiple commits on the same branch in a commit set, we now retry the CreateBranch transaction with a new transaction ID.