Stop pipeline during deletion. (#6522)
In some of our tests, we delete a pipeline which has commits being
created very frequently. Deleting a pipeline output repo can end up
being a large transaction, which can lead to DeleteAll hanging. If we stop
a pipeline as part of deletion, we can avoid this issue.

Previously, ability to delete a pipeline did not imply ability to stop a pipeline,
as stopping counted as an update, and so required read access from input repos.
This didn't make much sense, so now starting and stopping are a separate 
type of pipeline op which only requires write access to the pipeline repo.