Remove context from transaction context.
Nested transactions have the potential to cause deadlocks if they
conflict - the database might allow the inner transaction, not realizing
that when the outer one fails, the inner will be performed again. It's
also not a good idea in general to call a gRPC inside a transaction, in
case it involves waiting on some other process.

To make these operations hard to perform, the transaction shouldn't have
easy access to a context. Most existing usages were trivial to avoid by
only allowing them in the parent call, which had a context. Inspect
pipeline with details and checking enterprise limits were two such cases
where we never actually needed the context part in a transaction.

The remaining case is pipeline spec files. While we've discussed some
workarounds, creating and updating a pipeline still requires both
creating a fileset and reading a file, which cannot currently be done
transactionally. These operations are provided through an interface which
essentially just fills in the original context.