Switch s3 sidecar pipeline detail fetch back to gRPC (#6650)
This doesn't actually need to be over gRPC, just a side effect due to how context metadata works.
All of our code checks the incoming metadata for auth info, expecting to have received it from a request.
The other cases where the internal path is used, auth is mostly irrelevant, and so we just put a username 
for tracking and any future permissions concerns, but here it's a weird mixture of neither being
a generic background request or coming from a user request. For clarify, it seemed simpler to just flip back
to a gRPC and use the real context.

Also run s3 tests with auth enabled by default since it's already implemented.