Create transaction "refresh" layer.
To finally prevent nested transactions, we need a solution for creating and accessing filesets. Instead of trying to make fileset operations possible within a transaction, we create a new FilesetManager interface which appears to allow reading and creating filesets, but in fact returns an error that causes the transaction to restart. In between transaction attempts it performs the desired fileset operations and makes sure any data is up to date.

In addition, move the "internal" methods to just be exposed methods on the api server for ease of use and discoverability.