Remove context from transaction context. (#6647)
Nested transactions have the potential to cause deadlocks if they
conflict - the database might allow the inner transaction, not realizing
that when the outer one fails, the inner will be performed again. It's
also not a good idea in general to call a gRPC inside a transaction, in
case it involves waiting on some other process.

To make these operations hard to perform, the transaction shouldn't have
easy access to a context. Most existing usages were trivial to avoid by
only allowing them in the parent call, which had a context. Inspect
pipeline with details and checking enterprise limits were two such cases
where we never actually needed the context part in a transaction.

To finally prevent nested transactions, we need a solution for creating and accessing filesets. Instead of trying to make fileset operations possible within a transaction, we create a new FilesetManager interface which appears to allow reading and creating filesets, but in fact returns an error that causes the transaction to restart. In between transaction attempts it performs the desired fileset operations and makes sure any data is up to date.