Make cron ticks atomic (#6593)
The cron tick logic was separately opening a new commit, making changes, 
and then closing it. If, due to this non-transactionality or a user's accidental
modification, the head commit was already open when attempting to
write the tick file, the commit was squashed. This was a translation from earlier 
code which just deleted the commit, but squashing can have unintended side
effects. By making the cron ticks atomic, we can avoid having to worry about 
this at all.