Use existing timestamp and commit set ID for pipeline info.
With the hash-based fileset manager, we need every attempt at creating
the pipeline info to result in identical data. Using the timestamp and
ID from the transaction context isn't enough if they are reset every
time a new fileset is added. We must also set those value exactly once
at the beginning of the overall transaction attempt, outside of the
fileset manager loop.