Fix pipeline access in worker and master. These operations are more subtle as they involve non-transactionally changing the state of a pipeline.
The worker only ever operates on the specific pipeline version in its job. If somehow the worker is still running while a new pipeline version exists, it will (correctly) not touch the new version.

The important change in the PPS master logic is around failing a pipeline when step fails. Steps are supposed to operate on the current version, and so aren't tied to a particular one. There's a possibility that a pipeline update could happen in between the step failure and actually failing the pipeline, but this should be rare, and failures here are probably important enough to surface to users anyway.