Change delete pipeline behavior.
Most importantly, do everything in a transaction. Still stop first
outside of the transaction to avoid needless conflicts.

Also, delete the meta and spec repos in the keep repo case, so that
there's no longer any indication that the repo was a pipeline. This
necessitated preventing creating a pipeline on top of an existing
repo, as otherwise the pipeline output would be unpredictably appended
to existing files without any errors.