Store pipeline specs in postgres. (#6735)
Storing pipeline specifications as files in PFS has been a source of complication in operations involving pipelines. While it fits well with pachyderm's provenance model, it means many pipeline operations could not be truly transactional, and required an additional lookup in many cases with the overhead associated with accessing the data in object storage.

This change keeps all important pipeline information in the postgres pipeline collection. In addition, we retain old versions of the pipeline in the collection as well, keyed by the pipeline name and the commit ID which introduced that pipeline version. This lets us remove all the logic around managing filesets in transactions. 

The new complication that is introduced is that questions about whether a pipeline exists or what the current pipeline version is are no longer answered by looking up a fixed key in the pipeline collection (i.e. the name). To find the current pipeline version, a caller must examine the spec repo and find the relevant commit. In some other places, such as workers, the code wants to operate on a specific version of the pipeline, which becomes a simpler lookup with the new code.

Also, move all pipeline deletion steps into a single transaction. The current cron pipeline test seems to be hitting a pathological error case which often causes the delete to hang, which should be investigated. For now, that test just cleans up after itself more quickly, avoiding the error condition (possibly too many jobs?).

In addition, `--keep-repo` now deletes the meta and spec repos, leaving the output as a normal "source" input repo.