Remove special case logic for pachctl commands in workers.
When we switched spouts to use pachctl, we needed to add handling to
ensure that the client new to connect via localhost. This was
accomplished through an environment variable checked for some commands.
Our solution for allowing auth, however, also created an entire
pachyderm config via a secret.

This inadvertently made it difficult to override how a pipeline
connected to pachd, since the alternate path did not make use of the
usual environment variables. By making the special in-worker config
available to all workers and removing the special logic, the default
logic will connect to the sidecar as desired, without any surprising
behavior.