Reduce pipeline event load. (#6796)
A busy pachyderm cluster can generate enough pipeline modification
events to cause the controller to get backed up, introducing latency to
kubernetes state changes and possibly taking down pachd if the event
buffer is filled.

The main culprit is that every job state change writes
to the corresponding pipeline, despite nothing about the state of the
pipeline having changed. Luckily, the data that tracks jobs is purely descriptive and has no
bearing on pachyderm logic. It's also simple to generate, requiring only
an iteration through a pipeline's jobs. I've removed the job counts by
status map, filling in only the most recent job state, as that seems
more valuable. It is easy to add later.

A secondary issue is that, once event processing gets backed up, it may
end up processing several past events for a single pipeline. The
controller always acts on the most recent pipeline state, so these
subsequent updates are unnecessary. To avoid the wasted effort, events
now include the time they were generated, so that old ones can be
ignored, although the logic to do so isn't included in this PR.

Finally, a quirk of the standby logic means pipelines often move through
the RUNNING state unnecessarily, only to immediately move to STANDBY.
This no longer happens, and there is a test enforcing a lack of changes
to the pipeline replication controller during a stop or start.