Implement group and join input streaming (#6975)
Group and join inputs now iterate over filesets, avoiding the need for
everything to be in memory. The filesets are only scoped to the
iteration call, so repeated iteration will be less efficient than other
input types (in particular, a cross containing a join as an input other
than the first should be avoided).

For now the iteration order is essentially random, in the order of
hashes of the keys. The performance should be revisited, possibly with a
more sophisticated datum index computation function.

There's still the potential for unbounded memory use if many sub-datums
in a group or join share the same key. This tradeoff was made for
efficiency for now, and it's much harder to hit than the previous
limits, but it does mean that e.g. an empty join or group is still
not a drop in replacement for a cross.