Improve tracker GC query performance. (#7093)
In some database engines, NOT IN prevents certain optimizations (in
particular, treating the query as an anti-join) even when all columns
involved are marked NOT NULL. Switching this to an equivalent NOT EXISTS
query resulted in a dramatic speedup when tested on a customer cluster.

Additionally, to avoid needlessly holding a query open, add a limit to
the deletable object query and simply iterate over the results in a
slice. We already run the query multiple times until completion, so no
other code needs to change.