Allow incorrect use of update pipeline. (#7332)
With auth enabled, an update pipeline call results in a missing role
binding error as it tries to authorize a write on a repo that doesn't
exist. We detect spurious updates in the rest of the logic, and will
error in the case the repo already exists, so this auth check seems
unintended and doesn't add security.