Propagate validation errors to the meta commit.
The job registry relies on the meta commits for information about job
status and history, including whether another job is suitable as a
"base". Jobs which encountered validation errors should not be used as a
base, but previously the associated meta commit would not reflect this,
resulting in an error loop as it tried to wait for an already-errored
output commit.