[CORE-463] Make inputs failed message clearer (#7558)
Add a new terminal state (unrunnable) jobs which represents a job which cannot be run because its inputs are unavailable (e.g. their jobs have failed, or are themselves unrunnable). This updates the “inputs failed” error message to indicate that the job was unrunnable.  There are also a number of new tests.

*N.b.: end users will need to upgrade `pachctl` in order to interact with unrunnable jobs.*