debug: loki: this whole advancedTime check was unnecessary
The Loki server claims two invariants; it won't return logs that come after the end time, and it
uses an exclusive comparison between a log and the end time.  Since we set our end time to the
timestamp of the oldest log each iteration, this means we'll always make progress.  (Though we might
miss logs; if 2 logs have the timestamp and they appear at the chunk boundary, we'll unfortunately
skip the second one.)

If the server stops doing these things we're definitely in trouble though.