Fix failing test
It turns out that following the Loki logs incremented the start of the
window repeatedly, but never increased the _end_ of the window.  For
following logs, it really makes more sense to increase the end — or
leave it off entirely, as I do here.  In order for this to work, I use
zero time.Time values to indicate no start or end.

Future work would include increasing the start time to match the
most-recently-seen log.

This also changed the log line parsing, as Loki now returns a
structured log rather than the line itself.