Consult RC when un-crashing pipeline.
The missing secrets test was flaking due to the crashing monitor taking
it out of the crashing state prematurely. The logic was wrong in both
directions: it required the maximum number of pods to be present, even
if the RC requested fewer, and could include workers in its count which
were going to be terminated.

While it seems like checking the RC's status should be sufficient, I've
left the worker status check in place to capture e.g. a failure in
networking in the worker service which the worker pod itself doesn't
care about.