Docker size (#648)
* reduce docker image size

I was chatting with @molinamelendezj and he mentioned the container was a little large.  It looks like you're not removing any packages with make prune-deps, needs to be in the same instruction because of the copy on write overlay filesystem docker uses.  Feel free to close if it's not helpful

* pr feedback

* test

* test

* test

* test dockerignore

* adding type

* revert remove files

* cleanup

* remove dep

* remove python

* update message around canvas

Co-authored-by: Sean Slattery <sean.slattery203@gmail.com>