[CORE-744] Add a simple log wrapper to Loki stream entries (#7823)
This PR fixes a bug that breaks structured JSON logging when using pachctl logs and Loki. Some log messages written by components using Klog use their own format and are being forwarded as-is by Loki. We address this with a check to see if the entry is a JSON formatted string. If it is not, then we wrap it in a struct and serialize it.