[CORE-813] Consolidate the kube-event-tail and pgbouncer images
## Description
This PR repoints the `values.yaml` to images consolidated to the pachyderm user. In the future, when we want to update these images, the plan is to add a `dynamic config` section to our `.circleci` config and run the related steps if one of the images are updated. This way, an engineer can update a vendor image we point to, then run those steps on a push of their local branch.