Perform batch pipeline creation transactionally.
While users can create multiple pipelines in a single CLI command, we
would not execute them in a transaction by default. This is a particular
issue for multi-pipeline templates, where the leftover state prevents
the same command from being rerun (switching to update would still work,
but there's no easy to undo the partial change).

This change performs multi-pipeline requests in a batch transaction by
default. Single pipelines and multi-pipeline creation in an explicit
transaction still use the old behavior for simplicity.

To improve the error messages from batch transactions, transaction
requests now have a simple string form which should be enough to point
users to the right request.

Pipeline creation has an additional wrinkle where validation is
performed outside of a transaction due to API calls to kubernetes and
the enterprise server. This is now included in a preflight step in the
batch transaction handler, which enables secret validation for the
multi-pipeline use case as well as closes a loophole where transactions
could skirt enterprise limitations.