[CORE-708] Support YAML Connector Configs
## Description
This PR adds a `Config` field to `IDPConnector` objects so that they match the the `dex` documentation. Historically the `JsonConfig` field was use to configure the objects, the new `Config` field will be used going forward. `JsonConfig` is kept for backwards compatibility.

Due to a limitation with the [dexidp/dex library](https://github.com/dexidp/dex/blob/f49e7bc218ebffdc5bfb800c53c06d273958eef2/server/server.go#L562), we need to convert any YAML a user gives us back to JSON. `yaml.YAMLToJSON()` was used for this since it is a no op on existing JSON objects.

Whenever an `IDPConnector` object is created, the overall strategy is to check to see whether `Config` or `JsonConfig` is defined. In the case where both are defined, `Config` is defaulted to. Then the value is converted to JSON.