[CORE-894] Filter ListDatum results (#7986)
* Add a status filter to ListDatumRequest

This adds a simple filter to ListDatumRequest. The filter is an example datum message; it will match any datum which matches all of its members; a member matches if the datum member is any of the values of the example message. In other words, matching a member is an OR, while matching the example message across members is an AND.