Ignore no-op updates during branch propagation. (#7955)
Stopping and starting a pipeline could lead to new jobs in pipelines
downstream of that pipeline even if nothing actually changed. This is
one of the easier ways to get into an open meta commit state, and is
generally confusing behavior.

This change moves the check that a branch is consistent with its
provenance to the very beginning, ignoring unchanged branches entirely.
This means we can skip the check for subvenant branches (which
necessarily will have a new upstream commit).