[CORE-1066] Refactor PachctlBashCommand (#8330)
The existing `PachctlBashCommand `has implicit semantics: it only runs in a test; it can panic or fail the test; the config file is always stored under a name based on the test; state matters between runs.

This update adds new `PachctlBash` type which can be used to run scripts using `pachctl` in Bash. It can be run independently of a test; it always returns errors and never panics; the config file is stored in a caller-configurable path; state is preserved as long as the instance is not closed. A `Cmd` type which embeds `exec.Cmd` to provide easy access to buffered stdout/stdin is also provided.

This update does _not_ replace instances of `PachctlBashCmd`; rather it implements `PachctlBashCmd` in terms of `PachctlBash`.