leave pipelines in STANDBY until they're FINISHED
When monitorPipeline starts, the first thing it does is put its pipeline
in STANDBY[1]. Then, it listens on ciChan for commits in READY and takes
the pipeline out of STANDBY if it gets any[2]. Meanwhile, another goro
calls SubscribeCommit and puts any events into ciChan[3].

The issue is that the first goro discards PFS events where the commit is
FINISHING, so if the PPS master restarts while a commit is in FINISHING,
it'll come back up, put the pipeline in STANDBY, and then won't take it
out. This leaves the commit stuck in FINISHING until compaction finishes
much, much later.

[1]:
https://github.com/pachyderm/pachyderm/blob/8fca5b912358c15f67d8b19a32f84d2ba507571c/src/server/pps/server/monitor.go#L137
[2]:
https://github.com/pachyderm/pachyderm/blob/8fca5b912358c15f67d8b19a32f84d2ba507571c/src/server/pps/server/monitor.go#L163
[3]:
https://github.com/pachyderm/pachyderm/blob/8fca5b912358c15f67d8b19a32f84d2ba507571c/src/server/pps/server/monitor.go#L110