Migrate PutFileURL and GetFileURL to Go-Cloud-SDK (#8506)
To increase maintainability, this change migrates to using the go-cloud-sdk for interfacing with cloud storage providers. This PR addresses PutFileURL and GetFileURL and leaves the existing object client code in place for the storage instances. Future PRs will fully migrate to the go-cloud-sdk once we have battle tested it and are confident that using it does not introduce any strange bugs.