[CORE-1099] Address the issue of overlong Kubernetes resource names (#8559)
* Address the issue of overlong Kubernetes resource names

Replaces pipeline- with $PROJECT- for all projects; constrains project names to 51 characters.

The total allowed is 63 characters; no existing pipeline with a name
longer than 51 characters has ever been runnable (`pipeline-$NAME-v?`
only allows 51 characters of pipeline name and one character of
version).

This is not a complete solution but a workaround which prevents the addition of projects from exacerbating the existing naming issues. Future work is scheduled to address those.